refactor: group concentrated liquidity modules
