+++ apps/web/src/app/(networks)/(evm)/[chainId]/_ui/concentrated-liquidity-provider/index.tsx
@@ -33,7 +33,7 @@ import {
   type ConcentratedLiquidityState,
   concentratedLiquidityReducer,
   initialConcentratedLiquidityState,
-} from './concentrated-liquidity-state'
+} from './state'
 
 type Api = {
   onFieldAInput(typedValue: string, noLiquidity: boolean | undefined): void
