+++ apps/web/src/app/(networks)/(evm)/[chainId]/_ui/select-prices-widget/price-block.tsx
